Fix compilation warnings adding stricter flags
-std=c89 -pdeantic -Wextra -Werror -Wno-long-long
-Wno-missing-field-initializers

Also, fix some compiler warnings under windows. There's still a LNK4197
warning which is yet to be resolved

Reviewed-on: http://review.couchbase.org/26349
Tested-by: Mordechai Nunberg <mnunberg@haskalah.org>

+++ src/get.c
@@ -57,7 +57,7 @@ static int handle_single_key(pycbc_ConnectionObject *self,
             }
         } else {
             ttl = pycbc_IntAsUL(curval);
-            if (ttl == -1 && PyErr_Occurred()) {
+            if (ttl == (unsigned long)-1 && PyErr_Occurred()) {
                 PYCBC_EXC_WRAP_OBJ(PYCBC_EXC_ARGUMENTS,
                                    0,
                                    "Couldn't convert TTL",
@@ -115,7 +115,7 @@ get_common(pycbc_ConnectionObject *self,
 {
     int rv;
     int ii;
-    int ncmds = 0;
+    Py_ssize_t ncmds = 0;
     size_t cmdsize;
 
     pycbc_seqtype_t seqtype;
